requirements.txt
praw
facebook-sdk
